+++ src/index.html
@@ -12,9 +12,6 @@
 
     <body>
         <div id="root"></div>
-
-        <script src="https://unpkg.com/react@17.0.1/umd/react.development.js"></script>
-        <script src="https://unpkg.com/react-dom@17.0.1/umd/react-dom.development.js"></script>
         <script src="App.js"></script>
     </body>
